requirements.txt
# OCR Платформа - Зависимости
# ================================

# Основные библиотеки для веб-интерфейса
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0

# Обработка изображений и PDF
Pillow>=10.0.0
PyMuPDF>=1.23.0
opencv-python>=4.8.0
numpy>=1.24.0

# OCR движок
pytesseract>=0.3.10

# Утилиты
typing-extensions>=4.7.0
python-dateutil>=2.8.2

# Дополнительные зависимости для стабильной работы
watchdog>=3.0.0
tornado>=6.3.0
